+++ CHANGELOG.md
@@ -9,10 +9,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+<a name="v0.1.1"></a>
+## [v0.1.1] - 2020-02-23
+### Docs
+- **readme:** updated Contributor section
+
+### Feat
+- **makefile:** adding release task
+
+### Improvement
+- **gormlog:** now using defaultLogger for testing
+- **logger:** adding WithTimeFormat(...) option for root options and defaultLogger
+
+
 <a name="v0.1.0"></a>
 ## v0.1.0 - 2020-02-23
 ### Build
 - **makefile:** polish makefile
+
+### Chore
+- **changelog:** adding changelog
 
 ### Docs
 - **logger:** updated readme
@@ -40,4 +56,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **format:** fix code format
 
 
-[Unreleased]: https://github.com/xmlking/micro-starter-kit/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/xmlking/micro-starter-kit/compare/v0.1.1...HEAD
+[v0.1.1]: https://github.com/xmlking/micro-starter-kit/compare/v0.1.0...v0.1.1